changed int to floats

+++ Calculator/calculator_app.py
@@ -37,11 +37,11 @@ def calculator():
     """[summary]
     Do basic math functions with the ability to chain math functions on the previous result
     """
-    num1 = int(input("What is the first number?: "))
+    num1 = float(input("What is the first number?: "))
     for operation in operations.keys():
         print(operation)
     operation_symbol = input("Pick an operation from the line above: ")
-    num2 = int(input("What is the second number?: "))
+    num2 = float(input("What is the second number?: "))
     calculation_function = operations[operation_symbol]
     first_answer = calculation_function(num1, num2)
     print(f"{num1} {operation_symbol} {num2} = {first_answer}")
@@ -53,7 +53,7 @@ def calculator():
             f"Type 'y' to continue calculating with {first_answer}, or type 'n' to start a new calculation, or type 'e' to exit: ")
         if contin_calculations == "y":
             operation_symbol = input("Pick an another operation: ")
-            num3 = int(input("What is the second number?: "))
+            num3 = float(input("What is the second number?: "))
             calculation_function = operations[operation_symbol]
             second_answer = calculation_function(first_answer, num3)
             print(f"{first_answer} {operation_symbol} {num3} = {second_answer}")
